requirements.txt
streamlit>=1.30.0
openai>=1.10.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.24.0
